+++ jest.config.js
@@ -2,6 +2,10 @@ module.exports = {
   preset: "ts-jest",
   testEnvironment: "jsdom",
   moduleFileExtensions: ["ts", "tsx", "js", "jsx", "json", "node"],
+  moduleNameMapper: {
+    "^@/(.*)$": "<rootDir>/src/$1",
+    "^@form/(.*)$": "<rootDir>/src/app/form/$1",
+  },
   setupFilesAfterEnv: ["<rootDir>/jest-setup.ts", "jest-expect-message"],
   transform: {
     "^.+\\.(ts|tsx)$": [
