+++ src/app/form/(formSteps)/personal-details/1/PersonalDetailsStep1.test.tsx
@@ -1,8 +1,8 @@
+import PersonalDetailsStep1 from "@form/(formSteps)/personal-details/1/page";
+import { RouterPathnameProvider } from "@form/_utils/testUtils";
 import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { type AppRouterInstance } from "next/dist/shared/lib/app-router-context.shared-runtime";
-import { RouterPathnameProvider } from "../../../_utils/testUtils";
-import PersonalDetailsStep1 from "./page";
 
 const textInputFields = [
   { name: "First name *", key: "firstName", testValue: "Test first name" },
